refactor(core): move usePerspective useEffect to its own logic and provider
